backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Database
sqlalchemy==2.0.35
asyncpg==0.29.0
alembic==1.13.3
pgvector==0.3.3

# Redis / Celery
celery==5.4.0
redis==5.1.0
kombu==5.3.7

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
beautifulsoup4==4.12.3
anthropic>=0.40.0
slowapi==0.1.9

# OpenAI
openai==1.50.0
tiktoken==0.8.0

# Document parsing
pymupdf==1.24.11
python-docx==1.1.2

# HTTP client
httpx==0.27.2

# Validation / config
pydantic[email]==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0

# Monitoring
prometheus-fastapi-instrumentator==7.0.0

# Utilities
python-dateutil==2.9.0
bleach==6.1.0

# Dev/test
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
ruff==0.6.9
